Download function and test json option
With the new download function, we can
add an hidden option to load a json files
from disk instead of making an API call.

The option is hidden since it should
only be used in development.

It should be easier to test the script.

Using a download function will also make it easier
to handle erros with downloads.