`echo()` partial instead of `click.secho()`
`click.secho()` will not output colors if it's not in an interactive
terminal.

The `color` boolean can be set to override this function.
Forcing to always or never print colors.

Since we want to replicate this behaviour with an `color` option
for the `f1next` command, we can set a `partial` to define
in a single place the value of `color` for every single `click.echo`
and `click.secho()` call.

As an initial effort, this commit sets the `echo()` partial,
and replaces every `click.echo()` and `click.secho()` call with it.